Move `compute_object_lifetime_bound` into submodule `dyn_trait`
